# docs/openapi.yaml
openapi: 3.1.0
info:
  title: 隐私邮箱外部取件 API v2
  version: 2.0.0
  description: |
    本文档主要描述外部接口，并记录升级所需的少量管理兼容接口。服务提供两种
    取件能力：标准只读 IMAPS，以及使用 API Key 或派生取码 URL 读取验证码历史。

    管理端可为每个主号配置上游隐式 TLS IMAP 主机和端口；创建时默认使用
    `imap.mail.me.com:993`。修改规范化后的主机、端口或用户名代表切换邮箱来源，会清理
    旧来源的同步游标、兼容快照、消费历史、待处理 `\\Seen` 任务、v2 归档和 OTP 历史，
    轮换公开 IMAPS 的 `UIDVALIDITY`，并将同步状态重置为 `pending`。

    IMAPS 用户名是隐私邮箱地址。客户端可以使用本服务签发的 IMAP 密码登录，
    或通过 `POST /oauth2/v2.0/token` 获取一小时访问令牌后使用 XOAUTH2。
    邮箱只有只读 INBOX，支持 LIST、STATUS、SELECT/EXAMINE、SEARCH、
    FETCH/UID FETCH 和 IDLE；所有写操作返回只读错误。

    新 v2 alias 的 API Key、IMAP 密码、client ID、refresh token 与 access token
    均由本服务签发。legacy alias 的兼容轮换只替换 API Key 和由其派生的旧直达链接，
    保留消费记录、邮件快照和 IMAP Seen 状态；v2 凭证包轮换才会同时使旧 API Key、
    取码 URL、IMAP 密码、refresh token 和尚未过期的 access token 失效。
    v2 的只读 OTP URL 与消费 recent-mail URL 使用不同用途的派生 token，不能互换；
    原 v1 直达 token 只继续授权 legacy alias 的 `/api/v1/mail/recent`。
    迁移前已经领取的 legacy alias 不会因 schema 升级而轮换。兼容接口
    `/api/v1/aliases`、`/api/v1/mail/latest` 和 `/api/v1/mail/recent`
    保留原有路径、鉴权和响应语义。

    迁移前已经领取的 alias 保持 `legacy` API Key；新 alias，以及迁移前尚未领取
    pending Key 的 alias，才会补齐 v2 凭证。`latest_messages`、消费记录、IMAP `Seen`
    任务、pending Key 和 `api_key_prefix` 在迁移中保留。

    管理兼容接口同时挂载在安装级随机 `<admin-path>/api/v1` 和固定
    `/admin/api/v1`。固定入口仅用于旧客户端兼容，不提供管理界面；两个入口执行
    相同的登录限流、会话认证和 CSRF 校验。随机路径不是访问控制边界。
servers:
  - url: https://HOST
  # Admin operations below override this server with the installation-specific
...
paths:
  /api/v1/aliases:
    post:
      tags:
        - Legacy Compatibility
      operationId: createLegacyAlias
      summary: 兼容登记已存在的隐私邮箱
      description: |
        保留 v1 的字段拼写 `add_hide_my_eamil`、`icloud`。两个字段可放在 query、
        `application/x-www-form-urlencoded` 正文，或分别放在两处；query 与正文合并后
        每个字段必须恰好出现一次，且不能包含其他字段。正文为空时不要求
        `Content-Type`。使用部署级 `ICLOUD_API_OAUTH_TOKEN` 的严格单一 Bearer 鉴权；
        成功响应直接返回 `api_key` 和 `mail_api_direct_link`；v2 凭证随后仍可由
        已认证管理端读取。该接口只登记 Apple 端已经存在的隐私邮箱，不负责创建
        地址。query 与正文合计不得超过 4 KiB。
      x-runtime-constraint:
        query-form-merge: >-
          OpenAPI 无法表达跨 query 与表单正文的联合必填约束；服务端合并两处参数后，
          要求 add_hide_my_eamil 与 icloud 各恰好出现一次，并拒绝任何其他字段。
      security:
        - externalOAuthToken: []
      parameters:
        - name: add_hide_my_eamil
          in: query
          required: false
          description: 可替代表单字段；合并 query 与正文后必须恰好出现一次。
          schema:
            type: string
            format: email
        - name: icloud
          in: query
          required: false
          description: 可替代表单字段；合并 query 与正文后必须恰好出现一次。
          schema:
            type: string
            format: email
      requestBody:
        required: false
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/LegacyAliasRequest"
            example:
              add_hide_my_eamil: alias@icloud.com
              icloud: primary@icloud.com
      responses:
        "201":
          description: 登记成功并返回兼容 API Key 与直达链接
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegacyAliasResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/InvalidOAuthToken"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "413":
          $ref: "#/components/responses/RequestTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/Unavailable"
  /api/v1/mail/latest:
    get:
      tags:
        - Legacy Compatibility
      operationId: getLegacyLatestMail
      summary: 兼容读取当前最新完整邮件快照
      description: |
        使用原有 `Authorization: Bearer <API Key>`。请求可重复执行，不消费邮件，
        不写入消费记录，也不改变 IMAP `Seen` 状态；已领取 legacy Key 和 v2 Key
        均可按绑定 alias 读取。
      security:
        - aliasAPIKey: []
      responses:
        "200":
          description: 当前最新邮件快照
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegacyLatestMailResponse"
        "401":
          $ref: "#/components/responses/InvalidCredential"
        "404":
          $ref: "#/components/responses/MailNotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/Unavailable"
  /api/v1/mail/recent:
    get:
      tags:
        - Legacy Compatibility
      operationId: getLegacyRecentMail
      summary: 兼容消费最近一小时内的当前邮件
      description: |
        使用原有 `api_key` 查询参数（只能出现一次），接受 API Key、legacy alias 的
        原 v1 直达 token，或 v2 alias 的 recent-mail 专用 token。成功后原子记录消费
        状态并异步写入主号 IMAP `Seen`；同一 UID 的后续请求继续返回
        `MAIL_NOT_FOUND`。旧 legacy 直达链接保持路径和 token 不变；OTP 专用 token
        不授权此接口。
      security:
        - aliasAPIKeyQuery: []
      parameters:
        - name: api_key
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 当前快照已消费并加入 IMAP Seen 队列
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegacyRecentMailResponse"
        "401":
          $ref: "#/components/responses/InvalidCredential"
        "404":
          $ref: "#/components/responses/MailNotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/Unavailable"
  /api/v1/otp:
    get:
      summary: 读取当前隐私邮箱的验证码历史
      description: |
        返回最新优先、持续累积的裸数组，每个邮箱最多 100 条。
        重复读取不会消费验证码，也不会修改邮件已读状态。没有验证码时返回 `200 []`。
        调用方必须二选一：使用 Bearer API Key，或传入管理端签发取码 URL 中的 OTP
        专用派生 token。recent-mail token 和原 v1 直达 token 不授权此接口。
      security:
        - aliasAPIKey: []
        - derivedToken: []
      parameters:
        - name: token
          in: query
          required: false
          description: 取码 URL 中的 OTP 专用派生 token；使用时不得同时发送 Authorization。
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: 验证码历史；没有记录时为一个空数组。
          headers:
            Cache-Control:
              schema:
                type: string
              example: no-store
          content:
            application/json:
              schema:
                type: array
                maxItems: 100
                items:
                  $ref: "#/components/schemas/OTPRecord"
              examples:
                records:
                  value:
                    - otp: "123456"
                      time: "2026-08-11T12:00:00+08:00"
                empty:
                  value: []
        "401":
          $ref: "#/components/responses/InvalidCredential"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/Unavailable"
...
  /oauth2/v2.0/token:
    post:
      summary: 用 refresh token 换取 IMAPS XOAUTH2 访问令牌
      description: |
        此接口是只读 IMAPS 的配套认证接口，不构成第三种取件方式。
        client ID 与 refresh token 都属于同一个隐私邮箱并由本服务签发。
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              additionalProperties: false
              required: [grant_type, client_id, refresh_token]
              properties:
                grant_type:
                  type: string
                  const: refresh_token
                client_id:
                  type: string
                refresh_token:
                  type: string
      responses:
        "200":
          description: 一小时有效的 Bearer access token。
          headers:
            Cache-Control:
              schema:
                type: string
              example: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
              example:
                access_token: ACCESS_TOKEN
                token_type: Bearer
                expires_in: 3600
        "400":
          $ref: "#/components/responses/OAuthInvalidRequest"
        "401":
          $ref: "#/components/responses/OAuthInvalidClient"
        "429":
          $ref: "#/components/responses/OAuthUnavailable"
        "503":
          $ref: "#/components/responses/OAuthUnavailable"
...
components:
  securitySchemes:
    externalOAuthToken:
      type: http
      scheme: bearer
      bearerFormat: deployment OAuth token
    aliasAPIKey:
      type: http
      scheme: bearer
      bearerFormat: service-issued API Key
    derivedToken:
      type: apiKey
      in: query
      name: token
      description: 仅授权只读 OTP 历史；不能用于消费 recent-mail。
    aliasAPIKeyQuery:
      type: apiKey
      in: query
      name: api_key
    adminSession:
      type: apiKey
      in: cookie
      name: icloud_admin_session
...
  schemas:
...
    OTPRecord:
      type: object
      additionalProperties: false
      required: [otp, time]
      properties:
        otp:
          type: string
          pattern: "^[0-9]{4,8}$"
        time:
          type: string
          format: date-time
    TokenResponse:
      type: object
      additionalProperties: false
      required: [access_token, token_type, expires_in]
      properties:
        access_token:
          type: string
        token_type:
          type: string
          const: Bearer
        expires_in:
          type: integer
          const: 3600
    LegacyAliasRequest:
      type: object
      additionalProperties: false
      description: >-
        正文中的可选部分；query 与正文合并后两个字段必须各出现一次。该跨位置约束
        无法由标准 OpenAPI Schema 单独表达，调用方和生成代码必须遵守 operation 上的
        x-runtime-constraint。
      properties:
        add_hide_my_eamil:
          type: string
          format: email
        icloud:
          type: string
          format: email
    LegacyAliasResponse:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          type: object
          additionalProperties: false
          required: [alias, icloud, api_key, mail_api_direct_link]
          properties:
            alias:
              type: string
              format: email
            icloud:
              type: string
              format: email
            api_key:
              type: string
            mail_api_direct_link:
              type: string
              format: uri-reference
              pattern: "^/api/v1/mail/recent\\?api_key=.*$"
    LegacyLatestMailResponse:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          type: object
          additionalProperties: true
          required: [alias, message, synced_at, stale]
          properties:
            alias:
              type: string
              format: email
            message:
              $ref: "#/components/schemas/LegacyMailMessage"
            synced_at:
              type: string
              format: date-time
            stale:
              type: boolean
    LegacyRecentMailResponse:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          type: object
          additionalProperties: false
          required: [address, subject, snippet, sent_at]
          properties:
            address:
              type: string
              format: email
            subject:
              type: string
            snippet:
              type: string
            sent_at:
              type: string
              format: date-time
    LegacyMailMessage:
      type: object
      additionalProperties: true
      required: [id, message_id, received_at, from, to, cc, subject, text, html, attachments, has_attachments, body_truncated]
      properties:
        id:
          type: string
        message_id:
          type: string
        received_at:
          type: string
          format: date-time
        sent_at:
          type:
            - string
            - "null"
          format: date-time
        from:
          type: array
          items:
            $ref: "#/components/schemas/MailAddress"
        to:
          type: array
          items:
            $ref: "#/components/schemas/MailAddress"
        cc:
          type: array
          items:
            $ref: "#/components/schemas/MailAddress"
        subject:
          type: string
        text:
          type: string
        html:
          type: string
        attachments:
          type: array
          items:
            type: object
            additionalProperties: true
        has_attachments:
          type: boolean
        body_truncated:
          type: boolean
    MailAddress:
      type: object
      additionalProperties: true
      required: [email]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
...
    APIErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, request_id]
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
  responses:
...
    InvalidRequest:
      description: 参数缺失、重复、格式错误，或包含不支持的字段（`INVALID_REQUEST`）。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIErrorEnvelope"
    InvalidCredential:
      description: Bearer 凭据、API Key 或派生 token 无效。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIErrorEnvelope"
    InvalidOAuthToken:
      description: 部署级 OAuth Bearer 凭据无效（`INVALID_OAUTH_TOKEN`）。
      headers:
        WWW-Authenticate:
          description: Bearer 鉴权提示。
          schema:
            type: string
            const: Bearer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIErrorEnvelope"
    MailNotFound:
      description: 当前没有符合条件且尚未消费的邮件（`MAIL_NOT_FOUND`）。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIErrorEnvelope"
    NotFound:
      description: 请求的主号、alias 或邮件不存在。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIErrorEnvelope"
    Conflict:
      description: alias 已存在或主号已达到启用数量上限。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIErrorEnvelope"
    RequestTooLarge:
      description: query 与正文合计超过 4 KiB（`REQUEST_TOO_LARGE`）。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIErrorEnvelope"
    UnsupportedMediaType:
      description: 非空正文没有使用 `application/x-www-form-urlencoded`（`UNSUPPORTED_MEDIA_TYPE`）。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIErrorEnvelope"
    RateLimited:
      description: 请求过于频繁。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIErrorEnvelope"
    Unavailable:
      description: 服务暂不可用。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIErrorEnvelope"
    InternalError:
      description: 服务处理凭据或直达链接时发生内部错误。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIErrorEnvelope"
...
    OAuthInvalidRequest:
      description: OAuth 表单格式错误。
      content:
        application/json:
          schema:
            type: object
            required: [error, error_description]
            properties:
              error:
                const: invalid_request
              error_description:
                type: string
    OAuthInvalidClient:
      description: client ID 或 refresh token 无效。
      content:
        application/json:
          schema:
            type: object
            required: [error, error_description]
            properties:
              error:
                const: invalid_client
              error_description:
                type: string
    OAuthUnavailable:
      description: 令牌服务暂不可用或请求受到限流。
      content:
        application/json:
          schema:
            type: object
            required: [error, error_description]
            properties:
              error:
                const: temporarily_unavailable
              error_description:
                type: string